+++ packages/mint-components/src/components.d.ts
@@ -2027,6 +2027,11 @@ export namespace Components {
           * @uiName Last name label
          */
         "lastNameLabel": string;
+        /**
+          * @uiName Optional
+          * @uiWidget
+         */
+        "optional": boolean;
     }
     interface SqmNavigationMenu {
         /**
@@ -9733,6 +9738,11 @@ declare namespace LocalJSX {
           * @uiName Last name label
          */
         "lastNameLabel"?: string;
+        /**
+          * @uiName Optional
+          * @uiWidget
+         */
+        "optional"?: boolean;
     }
     interface SqmNavigationMenu {
         /**