+++ packages/mint-components/src/components/sqm-name-fields/sqm-name-fields.tsx
@@ -31,6 +31,12 @@ export class NameFields {
   @Prop() lastNameLabel: string = "Last Name";
 
   /**
+   * @uiName Optional
+   * @uiWidget
+   */
+  @Prop() optional: boolean = false;
+
+  /**
    * @undocumented
    * @uiType object
    */
@@ -51,6 +57,7 @@ function useNameFieldsDemo(props: NameFields): Partial<NameFieldsViewProps> {
   return deepmerge(
     {
       states: {
+        optional: props.optional,
         validationErrors: [],
         content: {
           firstNameLabel: props.firstNameLabel,