+++ packages/mint-components/CHANGELOG.md
@@ -7,7 +7,34 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.15.3] - 2025-11-14
+## [1.15.5] - 2025-10-21
+
+### Fixed
+
+- \<sqm-portal-google-registration-form>
+  - Fix terms checkbox slot to show on both form steps
+  
+## [1.15.4] - 2025-10-20
+
+### Added
+
+- \<sqm-portal-google-registration-form>
+  - Enhanced two-step registration flow with improved validation
+  - Added support for additional form fields in registration process
+
+- \<sqm-name-fields>
+  - Added optional prop
+
+### Updated
+
+- \<sqm-password-field>
+- \<sqm-marketing-emails-checkbox>
+- \<sqm-input-field>
+- \<sqm-dropdown-field>
+- \<sqm-checkbox-field>
+  - Added google registration form to valid parents
+
+## [1.15.3] - 2025-10-14
 
 ### Added
 
@@ -766,7 +793,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - \<sqm-portal-change-password>
   - \<sqm-portal-email-verification>
   - \<sqm-portal-forgot-password>
-  - \<sqm-portal-frame>
   - \<sqm-portal-login>
   - \<sqm-portal-logout>
   - \<sqm-portal-profile>
@@ -1289,7 +1315,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - \<sqm-popup-container>
   - \<sqm-stencilbook>
 
-[unreleased]: https://github.com/saasquatch/program-tools/compare/mint-components@1.15.3...HEAD
+[unreleased]: https://github.com/saasquatch/program-tools/compare/mint-components@1.15.4...HEAD
+[1.15.4]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.15.4
 [1.15.3]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.15.3
 [1.15.2]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.15.2
 [1.15.1]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.15.1